2nd added and fully tested (at first manually) /subscribe feature with custom bitcoin price

+++ src/main/java/com/skillbox/cryptobot/client/entity/Subscribers.java
@@ -38,7 +38,7 @@ public class Subscribers {
     Long telegramUserId;
 
     @Column(name = "user_subscribed_bitcoin_price")
-    Long userSubscribedBitCoinPrice;
+    Double userSubscribedBitCoinPrice;
 
     @Override
     public final boolean equals(Object o) {
